*) Code from Metty (Uthgard) - Bugfix: Fixed a deadlock when trading and crafting at the same time. *) Code from Lohx - Added: An experimental Dirty Tricks handler. *) Code from Me - Bugfix: Fixed an issue which prevented the Region loot generators from being assigned to a GameNPC. - Bugfix: All housing NPC's are given the peace flag, this makes them usable whilst we only have the Hibernian DB. - Added: The final set of support for PvP keep systems. - Bugfix: Mastery of Parrying and Mastery of Blocking RA bonuses are now taken into account with the calculator.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/branch/1.x@363 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/gameutils/LootMgr.cs
@@ -298,12 +298,14 @@ namespace DOL.GS
 
 			IList nameGenerators = (IList)m_mobNameGenerators[mob.Name];
 			IList guildGenerators = (IList)m_mobGuildGenerators[mob.GuildName];
-			IList regionGenerators = (IList)m_mobRegionGenerators[mob.CurrentRegionID];
+			IList regionGenerators = (IList)m_mobRegionGenerators[(int)mob.CurrentRegionID];
+
 			//IList factionGenerators = m_mobFactionGenerators[mob.Faction]; not implemented
 
 			ArrayList allGenerators = new ArrayList();
 
 			allGenerators.AddRange(m_globalGenerators);
+
 			if (nameGenerators != null)
 				allGenerators.AddRange(nameGenerators);
 			if (guildGenerators != null)
